# PERF: RangeIndex.append with the same index