```python
In [1]: import pandas as pd, numpy as np

In [2]: idx = pd.RangeIndex(1000)

In [3]: %timeit idx.append([idx] * 100_000)
362 ms ± 4.39 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # main
342 ms ± 4.07 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # pr
```